Refactoring: GeoDataRiverSurvey::RenameRiverPathPointCommand is moved to separated files.
