+++ libs/geodata/riversurvey/riversurvey.pro
@@ -154,6 +154,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.h \
            private/geodatariversurvey_moveriverpathpointcommand.h \
            private/geodatariversurvey_removeextensioncommand.h \
+           private/geodatariversurvey_renameriverpathpointcommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
@@ -209,6 +210,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.cpp \
            private/geodatariversurvey_moveriverpathpointcommand.cpp \
            private/geodatariversurvey_removeextensioncommand.cpp \
+           private/geodatariversurvey_renameriverpathpointcommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.cpp \
            private/geodatariversurveycrosssectionwindow_riversurveytabledelegate.cpp \
